lib: fixed arithmetic errors

+++ lib/fmctdc-lib-math.c
@@ -38,9 +38,9 @@ uint64_t fmctdc_ts_ps(struct fmctdc_time *a)
 {
 	uint64_t ps = 0;
 
-	ps += a->seconds * 1000000000000ULL;
-	ps += a->coarse * 8000;
-	ps += a->frac * 8000ULL / 4096ULL;
+	ps += (uint64_t) a->seconds * 1000000000000ULL;
+	ps += (uint64_t) a->coarse * 8000ULL;
+	ps += (uint64_t) a->frac * 8000ULL / 4096ULL;
 	return ps;
 }
